Update 3
I realized that, somehow, my code was duplicated. I'm not sure how this happened, but I tried to fix it by removing the part of the code that didn't reflect the changes I made in VS. I'm not sure if this was the wrong or right thing to do, but hopefully I'll be able to fix everything with help from the tutor I am meeting with tomorrow.